% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main function.R
\name{hdpcluster_ds}
\alias{hdpcluster_ds}
\title{Inference in high-dimensional data after discretizing unobserved heterogeneity}
\usage{
hdpcluster_ds(
  y,
  D,
  X,
  T,
  groups_unit = NULL,
  index,
  data,
  cluster_type = "unit kmeans",
  pesudo_type = "average",
  link = "average",
  optimal_index = NULL
)
}
\arguments{
\item{y}{outcome variable}

\item{D}{treatment variable}

\item{X}{control variables}

\item{T}{panel size of the time period}

\item{index}{index name}

\item{data}{data which contains the correct index of the outcome variable or the treatment variable}

\item{cluster_type}{different cluster means for the data, 'unit kmeans' is default, allow for 'unit pesudo'}

\item{pesudo_type}{if cluster_type = 'unit pesudo', choice of the pesudo type}

\item{link}{if cluster_type = 'unit pesudo', choice of different links of pesudo type}

\item{optimal_index}{if cluster_type = 'unit pesudo', different ways to compute the optimal number of clusters}

\item{groups_init}{number of the unit clusters}
}
\value{
A list of fitted results is returned.
Within this outputted list, the following elements can be found:
\item{res}{regression model.}
\item{G}{number of unit clusters.}
\item{estimate_correct}{corrected standard error, t value, and p value.}
\item{summary_table}{summary of the model together with corrected estimates in the 'Coefficients'.}
}
\description{
R package 'HDpcluster' is dedicated to do inference for high-dimensional linear panel data model with unkown functions of fixed effects.
}
\details{
Estimator Function with Optional Cross-Fitting
}
